% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRCAtlas_shiny.R, R/app.R
\name{.plotLR}
\alias{.plotLR}
\title{Scatter Plot of dose response curve}
\usage{
.plotLR(dose, response, n_para)

.plotLR(dose, response, n_para)
}
\arguments{
\item{dose}{"ligand -> receptor"}

\item{response}{gem name}

\item{n_para}{number of parameters in fitting nonlinear sigmoid function}
}
\description{
Scatter Plot of dose response curve

Scatter Plot of dose response curve
}
